fixed unsafety
turns out there are rules about what you can do with a pointer while a &mut or a &
to the same value exists. mainly not accessing/not modifying through it respectively.

this is burried deeep deep inside the docs, the only place i could find it was the
UnsafeCell doc (which does not give you a way around this constraint either)

so i had to abandon the regular iterator interface for a very similar custom implementation.